# apps/api/files/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Bitcoin API",
    "description": "Real-time Bitcoin price API.",
    "version": "0.0.1"
  },
...
  "paths": {
...
    "/api/v1/prices/current": {
      "get": {
        "operationId": "getCurrentPrice",
        "summary": "Get current price",
        "tags": [
          "prices"
        ],
        "description": "Returns the latest price for a given trading pair from the configured exchange. For websocket use: wscat -c ws://localhost:8000/api/v1/prices/current?symbol=btcusdt",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "btcusdt"
              ]
            },
            "in": "query",
            "name": "symbol",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "price": {
                      "type": "string"
                    },
                    "time": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "price",
                    "time"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
...
}